+++ app/dashboard/src/App.tsx
@@ -80,21 +80,22 @@ import * as suspense from '#/components/Suspense'
 
 import AboutModal from '#/modals/AboutModal'
 import { AgreementsModal } from '#/modals/AgreementsModal'
-import * as setOrganizationNameModal from '#/modals/SetOrganizationNameModal'
+import { SetupOrganizationAfterSubscribe } from '#/modals/SetupOrganizationAfterSubscribe'
 
 import LocalBackend from '#/services/LocalBackend'
 import ProjectManager, * as projectManager from '#/services/ProjectManager'
 import RemoteBackend from '#/services/RemoteBackend'
 
+import { FeatureFlagsProvider } from '#/providers/FeatureFlagsProvider'
 import * as appBaseUrl from '#/utilities/appBaseUrl'
 import * as eventModule from '#/utilities/event'
 import LocalStorage from '#/utilities/LocalStorage'
 import * as object from '#/utilities/object'
 import { Path } from '#/utilities/path'
+import { STATIC_QUERY_OPTIONS } from '#/utilities/reactQuery'
 
 import { useInitAuthService } from '#/authentication/service'
 import { InvitedToOrganizationModal } from '#/modals/InvitedToOrganizationModal'
-import { FeatureFlagsProvider } from '#/providers/FeatureFlagsProvider'
 
 // ============================
 // === Global configuration ===
@@ -178,14 +179,8 @@ export default function App(props: AppProps) {
         supportsLocalBackend: props.supportsLocalBackend,
       },
     ] as const,
-    meta: { persist: false },
     networkMode: 'always',
-    staleTime: Infinity,
-    gcTime: Infinity,
-    refetchOnMount: false,
-    refetchInterval: false,
-    refetchOnReconnect: false,
-    refetchIntervalInBackground: false,
+    ...STATIC_QUERY_OPTIONS,
     behavior: {
       onFetch: ({ state }) => {
         const instance = state.data?.projectManagerInstance ?? null
@@ -432,7 +427,7 @@ function AppRouter(props: AppRouterProps) {
       <router.Route element={<authProvider.NotDeletedUserLayout />}>
         <router.Route element={<authProvider.ProtectedLayout />}>
           <router.Route element={<AgreementsModal />}>
-            <router.Route element={<setOrganizationNameModal.SetOrganizationNameModal />}>
+            <router.Route element={<SetupOrganizationAfterSubscribe />}>
               <router.Route element={<InvitedToOrganizationModal />}>
                 <router.Route element={<openAppWatcher.OpenAppWatcher />}>
                   <router.Route
@@ -543,7 +538,7 @@ function LocalBackendPathSynchronizer() {
   const localBackend = useLocalBackend()
   if (localBackend) {
     if (localRootDirectory != null) {
-      localBackend.rootPath = Path(localRootDirectory)
+      localBackend.setRootPath(Path(localRootDirectory))
     } else {
       localBackend.resetRootPath()
     }