+++ app/dashboard/src/components/AriaComponents/Form/components/types.ts
@@ -123,6 +123,7 @@ export interface FieldProps {
   readonly label?: React.ReactNode | undefined
   readonly description?: React.ReactNode | undefined
   readonly error?: React.ReactNode | undefined
+
   /**
    * Defines a string value that labels the current element.
    */