+++ engine/runtime-integration-tests/src/test/java/org/enso/interpreter/test/TestLogProviderOnClasspath.java
@@ -1,18 +1,21 @@
 package org.enso.interpreter.test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.hasItem;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.nullValue;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.ServiceLoader;
 import java.util.stream.Collectors;
+import org.enso.logging.service.logback.test.provider.TestLogProvider;
 import org.junit.Test;
 import org.slf4j.spi.SLF4JServiceProvider;
 
 /**
- * In the `runtime/Test` testing suite, {@link org.enso.logger.TestLogProvider} should be among the
- * logging providers, because it is explicitly chosen as the logging provider for the tests.
+ * In the `runtime/Test` testing suite, {@link TestLogProvider} should be among the logging
+ * providers, because it is explicitly chosen as the logging provider for the tests.
  */
 public class TestLogProviderOnClasspath {
   @Test
@@ -25,12 +28,14 @@ public class TestLogProviderOnClasspath {
     }
     List<String> providerNames =
         providers.stream().map(elem -> elem.getClass().getName()).collect(Collectors.toList());
-    assertThat(providerNames, hasItem("org.enso.logger.TestLogProvider"));
+    assertThat(
+        providerNames, hasItem("org.enso.logging.service.logback.test.provider.TestLogProvider"));
   }
 
   @Test
   public void testLogProviderIsInUnnamedModule() throws Exception {
-    var testLogProviderClass = Class.forName("org.enso.logger.TestLogProvider");
+    var testLogProviderClass =
+        Class.forName("org.enso.logging.service.logback.test.provider.TestLogProvider");
     var mod = testLogProviderClass.getModule();
     assertThat(mod, notNullValue());
     assertThat("Should be an unnamed module - with null name", mod.getName(), nullValue());