+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   the entire input is a filtering pattern (it is not interpreted as parts of
   code). After picking any suggestion with Tab or new button the mode is
   switched to "code editing", where visualization preview is displayed instead.
+  Also the component browser help is now displayed in the right-side dock panel.
 - [Drilldown for XML][10824]
 - [Fixed issue where switching edited widget with <kbd>tab</kbd> key did not
   updated actual code][10857]
@@ -107,6 +108,7 @@
   `Comparable.new`][10468]
 - [Added `dec` construction function for creating `Decimal`s.][10517]
 - [Added initial read support for SQLServer][10324]
+- [Upgraded SQLite to version 3.46.1.][10911]
 
 [10434]: https://github.com/enso-org/enso/pull/10434
 [10445]: https://github.com/enso-org/enso/pull/10445
@@ -115,6 +117,7 @@
 [10474]: https://github.com/enso-org/enso/pull/10474
 [10517]: https://github.com/enso-org/enso/pull/10517
 [10324]: https://github.com/enso-org/enso/pull/10324
+[10911]: https://github.com/enso-org/enso/pull/10911
 
 # Enso 2024.2